Support llvm-objcopy for ElfCompactSectionsStep
Summary:
NDK r23 and above only contain llvm-objcopy (and not GNU objcopy), so
it's important to support it. llvm-objcopy always recalculates layouts
itself, and it doesn't currently support the `--change-section-address`
option. Adding support for that option to llvm-objcopy would be a
significant undertaking given how it's structured
(https://bugs.llvm.org/show_bug.cgi?id=45217 discusses this).

As far as I can tell, ElfCompactSectionsStep is only using that option
to force objcopy to recalculate the layout, not because the actual
addresses being set are important. The addresses are only relevant to
the dynamic linker, and this step is only used to produce interface
libraries, which will never be used at runtime (they're only meant to be
input to the static linker). If we have an objcopy that recalculates
layout automatically (such as llvm-objcopy), we can just call it without
any arguments to compact the sections, which is what this diff adds
support for.

Reviewed By: andrewjcg

fbshipit-source-id: dfed7a983c3f394a85a9d5104e7ca6a09d00ec39