Ensure ElfSharedLibraryInterface scratch file is writable
Summary:
The symbol table scrubbing steps modify the scratch file in-place, so it
needs to be writable. The scratch file is output from objcopy, and
llvm-objcopy copies permissions from its input file to its output file.
When running on RE, input files aren't writable, so the output file was
not writable either, and scrubbing would fail. Explicitly ensure the
file is writable to prevent this.

MakeWritableStep and MostFiles.makeWritable are pure copy-paste-modify
jobs of MakeExecutableStep and MostFiles.makeExecutable (respectively),
but the logic is trivial enough that I didn't think it was worth trying
to factor things out.

Reviewed By: andrewjcg

fbshipit-source-id: 6c139af0a90bd38d08c478599b9cf22ad7a1495e