Keep track of the output size of slow targets
Summary: The output size is given to us as part of the `BuildRuleEvent.Finished` event, which fires after `BuildRuleExecutionEvent.Finished`. This commit keeps track of the targets that are currently the slowest and, if an event is received that contains the output size, stashes that away on the `SlowTarget` for future logging to Scuba.

fbshipit-source-id: 9656109cf008a46b670daa62f88eaeb45fcae602