Remove maven Resolver
Summary:
It is standalone `java_binary`, not invoked with `buck` command, so I guess it is not used by anyone.

Last maven-specific commit to the resolver was done several years ago.

Removing because it is not trivial to convert the test to Starlark, and Python DSL is to be removed in the following D29589292.

Reviewed By: lebentle

fbshipit-source-id: 53916b6fd5454e6f97501c2ae0d2b03ef5f49627