Add secondary dexes via asset directories without renaming
Summary:
Preserve module dex names used in metadata.txt when building app bundles.

In order for native multidexing support to work with module dexes, the dexes and metadata should use classes*.dex as dex names, but this should probably be fixed at the point where dexes and metadata are produced (NonPreDexedDexBuildable), and not in the binary/bundle buildable code.

Reviewed By: longinoa

fbshipit-source-id: 301366c8a2d00d9d2b9963bb9f80a4bfa3363732