Improve pdb generation detection logic in buck.
Summary:
pdbs are only sometimes created in windows link jobs. It's dependent on
whether the *last* DEBUG flag passed to the linker is DEBUG:NONE or not (and at
least one DEBUG flag is passed). We weren't detecting -DEBUG, only /DEBUG, and the
way the logic was implemented wouldn't detect if we passed /DEBUG then passed
/DEBUG:NONE afterwards. Fix both of these issues.

Reviewed By: jdonald

fbshipit-source-id: 9be94e169f6a835f1948b0f33e18861e5131c61d