maybeReadProtocol() rename to readProtocol()
Summary:
The first iteration in the d_api reader is to expect a protocol type.

Before we use to have the main loop in the `run()` method and that is why we did check if the protocol is set to avoid reading it in each iteration. But right now we have a separate method that processes events from stream: `processEvents(InputStream stream)` and that is why we could rename `maybeReadProtocol()` into `readProtocol()`

Reviewed By: bobyangyf

fbshipit-source-id: ba7c82aba9c15b2038a95a35c2f6d9d57ec710c0