Support of multiple clients over the same json protocol.
Summary:
In case we use JSON protocol, we could expect the situation that multiple tools reuse the same event's named pipe to communicate back to buck.
The use case is that buck invokes genrule that invokes tool1 and then tool1 invokes tool2 and so on. If tool1 is connected to buck over D_API the first symbols it sends to establish the protocol is "j" followed by EOL. Then tool1 communicates with N events in JSON format and then invokes tool2 that is also trying to establish D_API connection with buck. So tool2 sends "j" followed by EOL and then events related to tool2's execution and when finished returned to tool1's process that continues to send events to buck.

Reviewed By: bobyangyf

fbshipit-source-id: eba2efa032425e565ec5eacaf53bc7fe826419e8