Add LinuxTimeParser, a parser for GNU time
Summary:
This commit is the first towards the goal of collecting resource utilization for processes that Buck spawns throughout the course of the build. To do this, we can use the GNU and BSD time programs (/usr/bin/time), which use the `getrusage` libc function to observe the resource usage of a process after it has terminated. GNU time allows for users to specify an arbitrary format string that will be filled in with the process's statistics and written to a file upon termination of the process. This commit is a small parser for a simple key-value file format (newline-deliminated key: value) that GNU time will write to disk when a shell command completes.

The objective is to use `/usr/bin/time` to interpose on shell steps and collect information on their resource utilization.

fbshipit-source-id: a97436848e7cdb71bc58324dcc5b3583dd5fcd71