readFromStream() with max chars param.
Summary:
D_API protocol requires to read protocol type followed by EOL and then in case of JSON protocol to read message length followed by EOL and then the message body.

So, if we have unexpected data in the input stream that doesn't have any EOL symbols we would read and collect the read chars in the buffer. This could cause OOM.

This diff brings a restriction and short but to this approach. We would read no more than max number of symbols and if no EOL is there then we fail with `InvalidDownwardProtocolException`

Reviewed By: bobyangyf

fbshipit-source-id: 1267b68fd8dbdb59ffc6379194e65359454a9c75