tell infer it's running under buck
Summary:
Infer used to rely on env var `INFER_ARGS` containing `--buck` so as to know it's running under buck. This is no longer going to work in RE as env vars don't translate.

This is used in infer to clean up the output and make it suitable for caching.

This diff makes the flavour explicitly tell the infer invoked it's running under buck via a command line argument.

Reviewed By: cjhopman

fbshipit-source-id: 888a83fbcd5ed291281100e755a56e0894ae4dad