Support macros in ocamldep_flags
Summary: The things passed to ocamldep_flags in the OCaml rules are often preprocessors, e.g. `buck/platform009/build/supercaml/share/dotopam/default/lib/lwt_ppx/ppx.exe`. However, because they are added as strings, Buck has no idea they are files, and because they are files not targets, it is a kind of package boundary violation. To fix the macros we need to switch to make a proper `export_file` target and use `$(location ...)`. To make that work, `ocamldep_flags` needs to take macros, which this diff does.

Reviewed By: IanChilds

fbshipit-source-id: 9816dd932d97406c375e09d514a6cf2a636b1dd8