Fix buck crash when creative spelling `buck kill `and `buck killall`
Summary:
Some sub commands (kill and killall) are implemented in Python wrapper, when users type some creative spellings like 'illall', it will pass to Java logic, the sub comment will be corrected to 'killall', but the corresponding Java logic throws IllegalStateException exception trace stack to users.

We fixed this issue by checking if the sub command is implemented by Java logic, if not, we generate CmdLineException and this exception will be handled by the command line framework.

Reviewed By: mykola-semko

fbshipit-source-id: b54d8f1e4db35bb987be6f9d4a7494be7670c176