Use CxxPlatform resolved from input nodes for generating Xcode projects
Summary:
Fetches the CxxPlatform using proper ways for Xcode project generation, instead of relying on the totally unsafe `defaultCxxPlatform`.

Selects the platform based on the platform flavor of the app input node, so that when generating project for individual targets within the app we know what CxxPlatform the app's building on.

Uses the correct platform for writing Xcode targets' configs, so that we're setting the right SDKs for targets in Xcode.

fbshipit-source-id: 87fdf49398ab14689f8863a58d5d32a9018b8088