disable RE for CxxInferCaptureTransitiveRule
Summary:
CxxInferCaptureTransitiveRule is uncacheable (fixed recently) and should not run on RE as it manipulates absolute paths.

The absolute paths are a design bug (it dates back to more than 6-7 years); this diff is a workaround until we fix this.

The upshot is that this rule is pointless to RE as it basically just concatenates a list of outputs.

Reviewed By: martintrojer

fbshipit-source-id: 6bc19c494d92c729913e550d21d3bb07eef1e016