Put BuildConfig classes in a more sane place
Summary:
Right now, `BuildConfigFields` is found in the `coercer` package. That means that anything depending upon it drags in a huge amount of core Buck dependencies. Instead, consolidate "build config" classes into a `build_config` package.

This reduces the binary size of the executable that we created in the previous diff from ~90MB to ~8MB.

Reviewed By: mykola-semko

fbshipit-source-id: 023393958789883507ba56cdc6eeeb11773ce478