parserconfig: use the eden_thrift globber when experiments.eden_globber is set
Summary:
Since parser.skylark_glob_handler is set in the various in repo configuration
file, a Buck experiment cannot override them. Since the experiment is used for
rollout, let's check it directly in Buck and once the rollout is over, we can
change the repository configuration and remove this code.

Reviewed By: stepancheg

fbshipit-source-id: d2eaa24c3499269e0647e1038e45548cf3141fad