Unwrap all process delegates instead of only the first one
Summary:
Prior to the advent of ResourceMonitoringProcesseExecutor, there was only one other delegate launched process executor - the executor responsible for launching processes using the Downward API. Now, it's possible to have three levels of delegation for process executors: a Downward API executor can delegate to a ResourceMonitoringProcessExecutor, which in turn delegates to a DefaultProcessExecutor. DefaultProcessExecutor unwraps delegate launched processes in order to find its impl class (since it expects to be the last entry in the delegate chain), but it was previously only unwrapping delegates one time, which does not work when the delegate chain is longer than two delegates long.

This commit switches the if to a while, which walks the full delegate chain to find the impl class at the end.

Reviewed By: mykola-semko

fbshipit-source-id: 521afab2ae4efc7e8653423d0a9bb73a5366ae70