cli: ignore .nfs silly rename files
Summary:
On an NFS mount, if an opened file is removed, the NFS client will perform a
"silly rename" to keep the file around until it is closed, at which point it
will be deleted.

Buck shouldn't care about these files since they are implemention details of
the NFS client and do not contain any meaningful data.

Reviewed By: stepancheg

fbshipit-source-id: 459ad4a1950da809f2e1643ee3fdd6a150002fc7