Default aaptMode to aapt2 and disallow aapt1
Summary:
No one uses aapt1 anymore, but it's still the default, which is bad, because it means that most of the tests are testing the wrong thing.

Also, it's a bunch of cruft (removed in next diff).

Policy on broken tests is mostly just to do whatever it takes to fix them, since they're currently not doing anything useful anyway.

Reviewed By: jiawei-lyu

fbshipit-source-id: 82f1b090c5caac065ccf29313df56e9b80f87d1b