AndroidManifest should merge placeholders, not Aapt2Link
Summary:
(Sorry for the large diff, but because we are moving code from one part of the pipeline to another, it can't be split into logical chunks).

Right now, the setup for creating an app's AndroidManifest is a bit weird - the `AndroidManifest` rule creates a manifest that still has placeholders, and then the `Aapt2Link` step is responsible for filling in the placeholder entries.

Instead, we make `AndroidManifest` responsible for creating the full manifest (including filling in the placeholder entries), and `Aapt2Link` just does aapt2 stuff

Reviewed By: jiawei-lyu

fbshipit-source-id: a66ada7ec6adccb29b516a5f88cb1688f87e6dd1