Disable DELAYEDEXPANSION in all buck genrules for cmd.exe (windows)
Summary:
Delayed Expansion is a little-used cmd.exe feature that allows the exclamation point to behave similar to the percent sign in batch scripts, except that substitution is done at evaluation of the command line, rather than at parsing. Typically this is used inside a `for` loop to allow the loop to modify a variable and pick up that value inside the loop body.

This feature can be turned on/off:
* Locally in the script via `setlocal`
* Locally for the cmd.exe process via the `/v:[on|off]` options.
* By a registry key for ALL cmd.exe processes.
The windows default is off.

The exclamation point is not escaped by buck's utility functions to handle special characters. Moreover, doing so would potentially override the genrule's commands, which could include a `setlocal` command to modify this state. Rather than try to guess, it's better if buck ensures that cmd.exe is executed such that it behaves as the windows default, namely it should always add `/v:off` when invoking `cmd.exe`.

This showed up in the build noted in the Test Plan. `fbsource/tools/build_defs/yacc.bzl`contains an exclamation point in `_WIN_YACC_CLEANUP_REGEX`, which causes the command line to `:bison_flex_cleanup` to be corrupted.

This has a better rundown of this than MSDN. See the section on bugs: http://www.nms-rennweg.ksn.at/allmann/allgemeines/1/scripts%20kommandozeile/Windows%20CMD%20Shell%20Command%20Line%20Syntax/EnableDelayedExpansion%20_%20Windows%20CMD%20_%20SS64.com.html

Reviewed By: mzlee

fbshipit-source-id: f50882e1a21aa5b775805cd19a5d448ce0c536d4