use fallback AppleCxxPlatform when apple.toolchain_set resolves to empty
Summary:
This change provides a fallback to local AppleCxxPlatforms when the `apple.toolchain_set` build rule has no platforms defined. This behavior is required to support setting a toolchain build rule as default without migrating all of the build targets to use it immediately. This will allow for a configuration like:
```
apple_toolchain_set(
  name = "toolchains",
  apple_toolchains = select({
      "some-toolchain-config": [define_toolchains()],
      "DEFAULT": [],
  }),
)
```

Reviewed By: KapJI

fbshipit-source-id: 4bc86b391ed2ddcc441d79d11af5668805859523