Merge library AndroidManifests by dep traversal order
Summary:
Currently, all library manifests are sorted by path name before merging. This changes the sort order to BFS traversal of the manifests on the dependency graph. The manifest merger interprets priority of library manifests by the order given, and this can affect the output of the final merged manifest.

https://developer.android.com/studio/build/manage-manifests#merge_priorities

Reviewed By: IanChilds

fbshipit-source-id: 40d24ede9e13bf0bdd79f2c39b198cc89bc59025