augment Apple fallback platform with buckconfig settings
Summary: This diff adds to the functionality that provides a fallback platform for when a provider based platform evaluates to empty.  The platform flags provided by a `CxxPlatformProvider` are augmented in `CxxPlatformsProviderFactory.updateCxxPlatformsWithOptionsFromBuckConfig` after the platform has been created to add the platform specific buck flags. This logic will not only apply to a `ProviderBackedCxxPlatform` so we need to add these values to the fallback platform at instantiation.

Reviewed By: KapJI

fbshipit-source-id: e7f2b328bea614ece5edc1bee23ebdce54ddc860