Export AAR deps for buck project
Summary:
Export AndroidPrebuiltAar deps for `buck project`.

This enables better compatibility with IntelliJ and aligns the treatment of AARs between `buck project` with `buck build`. Currently, AARs effectively export their deps when being built, but their deps were ignored in project generation. We already had this enabled for PrebuiltJar, so this just extends the support.

We could copy the filtering from AndroidPrebuiltAarDescription, but it provides almost no savings in testing as most dependencies for an AAR are JavaLibraries, AARs, or JARs anyway.

Reviewed By: cute-jumper

fbshipit-source-id: 242999294ae1946068a52869e41e491ba53d2ab7