Use min-sdk from manifest entries as min-sdk for predexed d8 rules
Summary:
Setting the minimum sdk here will pass it to `D8Step` for each predexed `DexProducedFromJavaLibrary`. This is required to build some libraries that otherwise fail during dexing with `Error: MethodHandle.invoke and MethodHandle.invokeExact are only supported starting with Android O (--min-api 26)`

This will cause debug/predexed builds to pass the same `--min-api` to d8 as is currently set for release builds.

A side effect of this change is that rulekeys for #d8 flavored rules are no longer only determined by the underlying java library, but also by the minimum sdk used by the android binary that depends on them, which will increase the amount of cache space used by d8 rules.

We could configure the minimum sdk for individual libraries that need it and not diverge the cache for every library/min sdk combination, but it's probably not worth the added complexity of using a different configuration than release builds.

Reviewed By: WaseemTheDream

fbshipit-source-id: 3cf147bc2cc5219ccadfd5855165436020b97f60