Improve efficiency of rulekey calculation for AndroidPrebuiltAar when using interface desugaring
Summary:
Upate AndroidPrebuiltAarDescription to only use the new dependency calculation path for interface desugaring for the PrebuiltJar we extract from the AAR.

Only the PrebuiltJar needs JavaLibrary dependencies for desugaring. Giving these extra JavaLibrary targets to the AAR target was regressing the time it took to calculate rule keys.

This provides a sizable speed improvement measured over 5 runs of `buck targets` on a large project.

Comparing with the `desugar_interface_methods_in_prebuilt_jars` flag off, on before this change, and on after this change:
| Flag off | Flag on (base) | Flag on (with this fix) |
| 83.5645 | 99.555239 | 74.846324 |
| 78.424422 | 101.365537 | 79.023176 |
| 73.60019 | 96.268712 | 75.915863 |
| 74.867371 | 100.19556 | 75.863137 |
| 72.288366 | 94.040165 | 77.147507 |

Averaged:
| Flag off | Flag on (base) | Flag on (with this fix) |
| 76.5489698 | 98.2850426 | 76.5592014 |

Reviewed By: IanChilds

fbshipit-source-id: 0cc45ca57e9b754480b063b7ca9f5fd26948b95e