pass target_sdk_version when linking tests
Summary:
This diff propagates a targets `target_sdk_version` value to the linker when linking test targets. This was incorrectly left out as the linker type is a bundle instead of a dylib, although both are dynamically linked.

This will avoid the large amount of log messages we get when linking test targets using lld.

Reviewed By: blackm00n

fbshipit-source-id: 50b593df2a6e83696fd23cd0aba7be65bb03123f