Support KSP in Buck
Summary:
Adding KSP (Kotlin Symbol Processing) support for Buck.

[KSP](https://github.com/google/ksp) is a new annotation processing engine for Kotlin, it's meant to replace KAPT and it can run alongside it. For more context on our plan to improve build times using KSP see [this post](https://fb.workplace.com/groups/2840058936242210/permalink/3032290147019087/) and other updates in the [Kotlin build speed group](https://fb.workplace.com/groups/2840058936242210).

KSP plugins are defined in BUCK files just like Java annotation processors, but the processor name needs to start with "KSP:". Example:

```
java_annotation_processor(
    name = "ksp_ap",
    processor_class = "KSP:com.example.ap.kotlinap_ksp.AnnotationProcessorKotlinKspProvider",
    visibility = [
        "PUBLIC",
    ],
    deps = [
        ":ksp_ap-lib",
    ],
)

kotlin_library(
    name = "ksp_ap-lib",
    srcs = [...],
    resources = glob([
        "resources/**",
    ]),
    ...
)
```

To run a KSP plugin in a given Kotlin module, we need to add it as a "plugin" and include the KSP library jars as kotlin_compiler_plugins. Example:

```
kotlin_library(
    name = "ap_testing",
    srcs = glob([
        "*.kt",
        "*.java",
    ]),
    plugins = [
        "//com/example/ap/kotlinap_ksp:ksp_ap",
    ],
    kotlin_compiler_plugins = {
        "//com/example/ap/libs:kotlin-symbol-processing-api.jar": {},
        "//com/example/ap/libs:kotlin-symbol-processing-cmdline.jar": {},
    },
    deps = [
        "//com/example/ap/libs:kotlin-symbol-processing-api.jar",
        ...
    ],
    ...
)
```
(For more examples, see the test cases below)

Note: Due to the way the KSP compiler plugin currently works, it needs to run in a separate kotlinc invocation, before the main compilation is happening with any other regular KAPT annotation processors. This creates a **limitation where code that's generated by KAPT is not processed by KSP**, for a given Kotlin module. KSP engineers gave their blessing for running KSP as part of the main compilation, they didn't implement the support for this yet but gave us instructions how to add it (see T106108430). This limitation doesn't hurt us at the moment, but we'll work on these KSP changes soon to lift this limitation.

Reviewed By: rybalkinsd

fbshipit-source-id: 2eb1d193bffad63a1ea2c582e47c7fe6d1d7aae0