Support of Incremental Swift builds
Summary:
## This diff introduces support of incremental Swift builds in BUCK.

So far, BUCK invoked the Swift driver in Whole Module Optimisation mode, which forced the driver to emit one frontend job per module. With WMO after each change, the whole module was be recompiled.

With Incremental mode enabled, the driver compiles each source-file "individually" (even within a single frontend job) and emits swiftdeps file which describes dependencies between files in a module and between different modules and allows the driver to skip compilation of files which were not changed or do not depend on changed declarations.

(More context: https://fb.workplace.com/groups/1540925742908161/posts/1560764750924260)

To enable Incremental mode, BUCK needs to emit an OutputFileMap.json file and pass it to the Swift driver.
This file is emitted per module and describes where every intermediate output file should go.

For incremental mode, the output of the SwiftCompile rule will be several object files instead of one as it is with WMO mode.
However, the module's swiftmodule file will always be single. (Driver is responsible for merging partial swiftmodules)

"Cross Module Incremental Imports" are enabled by the "-enable-incremental-imports" flag additionally passed to the driver, but that doesn't change anything from the BUCK perspective.

fbshipit-source-id: afcdb736e1d648da3b04f008cf36417d76ec90e8