Prepare a bunch of buck2-specific targets for visibility
Summary:
buck2 doesn't currently validate visibility which means we have a
bunch of violations. To prepare for buck2 validating it, make a bunch of things
public.

This also requires changing a bunch of rules/macros to support visibility. A lot
of the config_backed stuff is updated to pass through any kwargs as I've hit this
sort of thing before with them. Other (mostly non-prelude) rules need to get a
visibility attr added (buck1 rules all have it in prelude/attributes.bzl).

Reviewed By: stepancheg

fbshipit-source-id: 0181eb06babe282491fa9ac9072243dafb05b908