Fix windows race condition when unpacking pex
Summary:
On windows that move operation intermittently fails. The assumption for
now is the culprit is either the anti-virus or a file indexer, even though
there are conflicting signals about it. The end result is the new folder
is there, while the old still exists. Do retry with exp. back-off to mitigate.

More context:
https://fb.workplace.com/groups/2176375462609842/posts/2962967273950653/
https://fb.workplace.com/groups/2176375462609842/permalink/3090468361200543/

fbshipit-source-id: 122cb7abdf1920490ad5fae83226fbfb6a142c2f