Add option to not add runtime_deps to deps
Summary:
There's a longstanding bug where we add the `runtime_deps` to the classpath, which kind of defeats the point of `runtime_deps` - since they should only be used at runtime and not be needed at compile time.

This adds a flag to let us turn off this behavior once we've eliminated the dependency on it from our codebase.

Reviewed By: mykola-semko

fbshipit-source-id: 6304c059aba21c80e73deb32d434dcd5f917f5c6