use cxxDeps for swiftmodule deps
Summary:
This changes the dependent swiftmodule logic to use the `cxxDeps` to get dependent swiftmodule files. This works because the `SwiftCompile` rules provide generated headers for the preprocessor, so will be contained in the header dependency rules.

This fixes an issue where updating the swiftmodule files after the `getBuildDeps()` call will not add them to the build dependencies, as the result is cached by this call.

Reviewed By: milend

fbshipit-source-id: d6c120dcc4ccec774e8d5c9f4131a22f7b3cb6f0