add sdkDependenciesPath to swift_toolchain
Summary:
This diff adds a new field to `SwiftPlatform` where we can store SDK dependency information for a given SDK and platform. This is a requirement for explicit swift compilation as we need to know the `exported_deps` of the SDK frameworks in order to trigger compilation jobs for the swiftinterface files in the correct order.

A script to generate the expected input is included in `scripts/generate_swift_sdk_dependencies.py`.

Reviewed By: milend

fbshipit-source-id: 093d308b418a7ac5b530f7062dbae485f04495c4