add SwiftInterfaceCompile rules
Summary:
This diff adds a new build rule to compile `.swiftinterface` files into `.swiftmodule` files. This is used to compile the SDK modules when using explicit modules.

The `SwiftSdkDependencies` class maintains a cache of the transitive dependencies for each SDK module. Each `SwiftPlatform` has an instance of this class, so it is fixed per compiler, target and SDK. The dependencies should be reusable across all compilation actions using this platform.

Reviewed By: milend

fbshipit-source-id: c53c95cc9ff7fe48055d3bca04b83779f4133549