wrap SDK dependencies in ModuleOutput class
Summary: Wrap the SDK dependencies that are cached for each module in a ModuleOutput class so that we know which kind of module this is. This simplifies the graph walking logic when mixing module output types and means we only need a single `LoadingCache` instead of having the complexity of tracking separate types in different caches.

Reviewed By: maxovtsin

fbshipit-source-id: a8d1fab96fa3f085d6c2f72f91bbeea45fe9256e