add ExplicitModuleInput
Summary: This change adds a wrapper class to keep the base SourcePath and relative path for a module compilation input together. This makes the logic simpler in the build rules as we don't need to pass around all possible path prefixes and spread the path fixing logic in multiple places.

Reviewed By: milend

fbshipit-source-id: 05ebb1a366e7f8964700792fc73ab2108787a3cc