Add metadata.txt to MultiDexExecutableMain
Summary: We need to create a `metadata.txt` file which has a single line per dex, consisting of the jar name, the hash of the jar, and a class from inside the jar. Right now, we don't implement the third of these, as we don't have an easy way to look them up. If it is indeed required, we can add that later on.

Reviewed By: mykola-semko

fbshipit-source-id: 32d6450af680c8283ee0b8eec787bc0d3d6dda63