Complicate CxxToolchainDescription enough to actually support Windows
Summary:
Assembly for Windows, at least as currently implemented
in buck1, is just funky enough that we have to add explicit support for
the full range of as/aspp/asm/asmpp, instead of just copying values
from cxx/cxxpp/etc. So extend CxxToolchainDescription to read the
extra values, if they exist, and keep the same defaults as before
if they don't.

Reviewed By: KapJI

fbshipit-source-id: 364882926d8e31fbda1018f1d167497657576ae6