Buck install records adb port for android installs
Summary:
Buck uses ADB (Android Device Bridge) to install apk and other files to android device/emulator.
With this change buck will report the port it uses to connect to ADB server.
Instances of Android OD will have a dedicated port to communicate to local ADB server.
We'll use that information as proxy to identify if it is a remote install (running `buck install` on OD instance to install on local device/emulator).

Reviewed By: christolliday

fbshipit-source-id: 055839054e7aae0f4208a51a13a56e2ecfdc7fa9