Making usage of adb.agent_port_base in buck config more convenient
Summary:
We set agent port base in buck config `adb.agent_port_base` which defaults to 2828.
However, the first port being used is 2828 + 1 = 2829 which is a bit confusing.
For on-demand we need to setup of reverse tunnel for that port to enable buck install into the device attached to laptop.
This change makes the port we need to tunnel be the same one as we have to provide to buck which is less error prune.

Reviewed By: christolliday

fbshipit-source-id: 3e0c9c0b8655f1ea537b67cc502d9631cc5522fa