Add logcat crash information to instrumentation test output
Summary:
When an instrumentation test has a hard crash (SIGSEGV, SIGABRT, etc.),
the test runner loses the connection to the test and simply reports
that the process exited and more information can be found in logcat.
(This message comes from ddmlib.)  Now, we detect this message,
pull down the log from logcat, and look for the last crash report
from debuggerd.  If we find it, we include it in the test output.

Right now, this behavior is hard-coded, but I could make it enabled
by a flag.  That would require a separate diff (and release process)
for our external test runner, though.

Reviewed By: asm89

fbshipit-source-id: 4e38496a748b6b46b6bfe65add1ecba75856ac47