Fix bug with lambdas created by inline functions
Summary:
- Initially reported by nicholeic in [this post](https://fb.workplace.com/groups/1435783126588872/posts/2092677677566077/).
- Root cause found in D34520079.

The issue initially seemed to be related to inline functions that created
`suspend` lambdas in their bodies.

- The `suspend` lambda is represented by a class which contains `synthetic`
  fields to hold the lambda's local variables, so that when `invokeSuspend`
  returns after suspending the lambda, it can stash the local variable to pick
  back up where it left off.
- Lambdas belonging to inline functions are duplicated in their calling context
  (a new  class is created for them).  This duplicate is based on the ABI that
  the caller is given for that lambda.
- Class ABI generation strips things that it doesn't deem relevant for the ABI,
  including synthetic fields, but is aware of Kotlin's inline functions, so
  retains relevant method bodies.
- This creates a problem where the lambda's body was retained but the fields it
  refers to were not.

For suspend lambdas, this causes a crash, which is what we noticed.  Interestingly,
`CalculateClassAbi` does the same things to lambdas with captures -- it strips the
variables that hold the captures, but somehow `kotlinc` is able to re-constitute
the necessary information in that case.  I didn't think it was wise to rely on
this behaviour, so I've made it so that `CalculateClassAbi` retains everything
for any lambda that is owned by an inline function.

Reviewed By: IanChilds

fbshipit-source-id: f667c8904675a8d038bd7f2735332b9b0cb1bb25