add all swiftmodule dependencies to linker args
Summary: This diff changes the way swiftmodule paths are propagated to the linker. Instead of passing in exported_linker_flags for each apple_library with Swift, we collect the transitive swiftmodule dependencies in a set and include it in the `NativeLinkableInput`. This allows for passing in SDK swiftmodule dependencies to the linker without duplicating their paths for each linked library.

Reviewed By: milend

fbshipit-source-id: 7e63ce060c2ec8a9fa63342a633cd4bc561a66e4